Fix copilot comments
