+++ platform-sdk/swirlds-virtualmap/src/testFixtures/java/com/swirlds/virtualmap/test/fixtures/sync/BlockingInputStream.java
@@ -3,6 +3,7 @@ package com.swirlds.virtualmap.test.fixtures.sync;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InterruptedIOException;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
@@ -43,7 +44,10 @@ public class BlockingInputStream extends InputStream {
      */
     @Override
     public int read() throws IOException {
-        while (locked.get() && !Thread.currentThread().isInterrupted()) {
+        while (locked.get()) {
+            if (Thread.currentThread().isInterrupted()) {
+                throw new InterruptedIOException();
+            }
             Thread.onSpinWait();
         }
         return in.read();
